Feat: purchaseLotto 페이지와 비동기 통신 구현
최신의 lottoRound를 가져오는 기능과 lotto를 주문하는 기능을 구현하였다.

+++ virtualLottoSimulator/src/main/java/com/virtuallotto/virtuallottosimulator/service/LottoMachine.java
@@ -19,8 +19,8 @@ public class LottoMachine {
 
     private final LottoService lottoService;
 
-    public List<Lotto> generateTickets(int purchaseAmount) {
-        int ticketAmount = purchaseAmount / NumberConstants.LOTTO_PRICE.getValue();
+    public List<Lotto> generateTickets(Long purchaseAmount) {
+        int ticketAmount = purchaseAmount.intValue() / NumberConstants.LOTTO_PRICE.getValue();
         return generateLottoNumberRepeatNTimes(ticketAmount);
     }
